Week26/Day1/drf_beginners/posts/views.py
from rest_framework import mixins, generics
from .models import Post
from .serializers import PostSerializer


class PostMixinListView(
    mixins.ListModelMixin, mixins.CreateModelMixin, generics.GenericAPIView
):
    queryset = Post.objects.all()
    serializer_class = PostSerializer

    def get(self, request, *args, **kwargs):
        return self.list(request, *args, **kwargs)

    def post(self, request, *args, **kwargs):
        return self.create(request, *args, **kwargs)
